pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-runtime"
version = "0.1.0"
description = "Hybrid HFSM + behavior tree + event-sourced blackboard runtime for agentic systems"
requires-python = ">=3.10"
dependencies = ["litellm>=1.0.0", "mcp>=1.0.0", "fastapi>=0.115.0", "uvicorn>=0.30.0", "python-dotenv>=1.0.1"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentic_runtime*"]
